requirements.txt
flask==3.0.2
Pillow==10.2.0
python-dotenv==1.0.1
werkzeug==3.0.1
google-generativeai==0.3.2 